// frontend-admin/src/api/generated/model/tenantDeleteDependenciesDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */
import type { TenantDeleteDependencyBlockerDto } from './tenantDeleteDependencyBlockerDto';
import type { TenantDeleteDependencyCountsDto } from './tenantDeleteDependencyCountsDto';

export interface TenantDeleteDependenciesDto {
  /** @nullable */
  blockingDependencies?: TenantDeleteDependencyBlockerDto[] | null;
  canHardDelete?: boolean;
  dependencies?: TenantDeleteDependencyCountsDto;
  /** @nullable */
  failureCode?: string | null;
  /** @nullable */
  failureMessage?: string | null;
  hasDependencies?: boolean;
  hasFiscalFootprint?: boolean;
  /** @nullable */
  nextSteps?: string[] | null;
  tenantId?: string;
  /** @nullable */
  tenantSlug?: string | null;
  /** @nullable */
  tenantStatus?: string | null;
}